requirements.txt
pandas == 2.2.2
streamlit == 1.37.1
pillow == 10.4.0
scikit-learn == 1.6.0
tensorflow == 2.18.0
numpy == 1.26.4
mpmath == 1.3.0
sphinxcontrib-jsmath == 1.0.1
pip == 25.0
opencv-python == 4.10.0
opencv-python-headless == 4.10.0